+++ include/shiokaze/surfacetracker/macsurfacetracker3_interface.h
@@ -47,7 +47,7 @@ public:
 	 @param[in] fluid Fluid level set.
 	 \~japanese @brief 流体と壁のレベルセットをセットする。
 	 */
-	virtual void assign( const array3<double> &solid, const array3<double> &fluid ) = 0;
+	virtual void assign( const array3<float> &solid, const array3<float> &fluid ) = 0;
 	/**
 	 \~english @brief Advect level set.
 	 @param[in] u Velocity with which to advect.
@@ -56,14 +56,14 @@ public:
 	 @param[in] u 移流に使用する速度場。
 	 @param[in] dt タイムステップサイズ。
 	 */
-	virtual void advect( const macarray3<double> &u, double dt ) = 0;
+	virtual void advect( const macarray3<float> &u, double dt ) = 0;
 	/**
 	 \~english @brief Get the level set.
 	 @param[out] fluid Output level set.
 	 \~japanese @brief レベルセットを取得する。
 	 @param[out] fluid 出力のレベルセット。
 	 */
-	virtual void get( array3<double> &fluid ) = 0;
+	virtual void get( array3<float> &fluid ) = 0;
 	/**
 	 \~english @brief Draw level set surface.
 	 @param[in] g Graphics engine.
